fix: address code review findings across core modules
Bug fixes
- batch_fx: est_pi_sig() returned string "nan" instead of np.nan, silently
  breaking the pi_0 > 1 guard in subset_svd(); update caller check to
  np.isnan()
- CalcP: params = p0 was outside the else block, causing NameError when a
  .pkl null distribution file is supplied
- pirs/simulations: 1 / score produced inf for zero-PIRS genes in the
  precision-recall curve; replaced with negation (-score)
- pipeline: assert + assert 0 replaced with a proper ValueError; nested
  except ValueError chain fixed so the fallback branch is reachable

Refactoring / code quality
- classify: _REF_DIR now uses importlib.resources instead of os.path so
  it is robust under symlinks and editable installs; _source normalized to
  pd.DataFrame in __init__ (loaded once, no repeated isinstance checks)
- limbr/_normalize.py: new shared module with _qnorm() and _pool_norm();
  batch_fx and old_fashioned both drop their duplicated inner definitions
- pipeline: duplicated noreps SD-estimation block extracted to
  _setup_noreps_args() helper
- limma_voom: blanket RuntimeWarning suppression replaced with conditional
  pre-filtering (compute nanstd only for rows with n >= 2)
- BooteJTK: __score_at_percentile__ renamed to _score_at_percentile
  (dunder reserved for Python special methods); dead commented-out
  get_SD_distr() removed
- arbfit: model function parameter p renamed to params for clarity
- get_stat_probs: arctanh clip constant 0.99 extracted to _TAU_CLIP
- pirs/rank: get_tpoints() raises a descriptive ValueError on malformed
  column names; document intent of silent all-zero row filtering; add
  forkserver context comment
- pirs/simulations: concat-in-loop replaced with list + single concat
- simulations: lam_miss type annotation int -> float; constitutive mean
  assumption documented
- limbr/imputation, old_fashioned: dead self.notdone attribute removed

Project
- pyproject.toml: ipykernel moved from mandatory deps to optional
  [interactive] group; --cov flags removed from addopts so plain
  poetry run pytest works without the dev dependency group

+++ circ/simulations.py
@@ -77,7 +77,7 @@ class simulate:
         pbatch: float = 0.5,
         effect_size: float = 2.0,
         p_miss: float = 0.0,
-        lam_miss: int = 5,
+        lam_miss: float = 5,
         rseed: int = 0,
     ) -> None:
         pconst = 1.0 - pcirc - plin
@@ -128,6 +128,8 @@ class simulate:
                 ]
                 raw[idx] = [v for t in zip(*reps) for v in t]
             else:  # constitutive
+                # Mean of 1 keeps the row non-zero so the unit-SD scaling denominator
+                # is well-defined after scale() is applied.
                 raw[idx] = np.random.normal(1, amp_noise, n_cols)
 
         self._raw = raw
